unity/draft#screen/refresh-rate-ratio to unity/unity#trunk
This PR fixes case 1405134 which is actually a long standing issue of us "guessing" refresh rate when using exclusive full screen on Windows. The reason for this was the fact that Screen.resolutions returned rounded refresh rates, and so Screen.SetResolution took the rounded values that the users would pass in. Then, we would try to convert that to fractional values to pass to DirectX to switch the display mode. That was awkward and error prone (as evidenced by case 1405134, which exposes a flaw in our rounding mechanism).

This PR changes public APIs that used refresh rate as integer to instead use the RefreshRate struct, which is has a "numerator" and a "denominator", making refresh rate be a rational number.

Practically, all platforms fell into three buckets:

1. Refresh rate is completely ignored and hardcoded to 60 or 0 (cloud rendering, embedded Linux, Switch, WebGL);
2. APIs like Screen.currentResolution try to return the actual refresh rate, but don't support changing the refresh rate through Screen.SetResolution (Android, iOS, Linux, macOS, PS5, Xbox, UWP);
3. Screen.currentResolution returns the real refresh rate and Screen.SetResolution actually changes the refresh rate (Windows when running in exclusive fullscreen, and I *believe* PS4).

For buckets 1 & 2, there isn't much change. For bucket 3, our API becomes closer to what actually gets set when you call Screen.SetResolution and specifically on Windows, allows finer control of the refresh rate.