[Backport][2022.3][VFX] Reimport existing shader graph that support VFX when needed
https://jira.unity3d.com/browse/UUM-12579

🎮To reproduce:
1. Open Unity and create the new "2D URP" project
2. Install the Visual Effect Graph from the Package Manager
3. In the Project View Right-Click and select Create > Visual Effects > Visual Effect Graph
4. Leave the default "New VFX" name and confirm it with the Enter button
5. Open the "New VFX.vfx" Asset in Visual Effect Graph Window
6. Under Output Particle Quad node, select Shadergraph
7. Ensure the filter has the "Show Packages results" field enabled to see the default VFX Shadergraph

 
👍Expected result
In the "Select VFX Shader Graph" window you can see the "VFXSpriteLit" and the "VFXSpriteUnlit" Shader Graphs
👎Actual result
In the "Select VFX Shader Graph" window you can't see Shader Graphs just the "None" option

🔸Notes:
I’ve only managed to select those VFX Shader Graphs (the "VFXSpriteLit" and the "VFXSpriteUnlit") when I’ve copied them into the Project Asset directory

🏷Reproducible in Unity
`2022.2.0a16`, `2022.2.0b6`, and `2023.1.0a8`

💬Developer notes:
The issue happens because these ShadeGraph assets were already imported when the Visual Effect package is installed. So VFX code do not re-import them. But when they were first imported, the needed `ShaderGraphVfxAsset` resource had not been generated (because the VFX package was not installed yet).

We already have a mechanism to re-import VFX assets when the package version change (which happens on package installation). So the fix rely on this mechanism and in addition to re-importing VFX assets we also re-import ShaderGraph (only those that are compatible with VFX).

The symptom of this issue can be checked looking at the Shader asset like in the screenshots below
<img width="158" alt="Unity_IHgBLIm0xt" src="https://media.github.cds.internal.unity3d.com/user/4003/files/b585ca15-49f6-44d6-93b9-a0d02eb3a374">
But we expect this:
<img width="186" alt="jZuk0ULmGx" src="https://media.github.cds.internal.unity3d.com/user/4003/files/49ae171e-5759-4273-9513-d9f930a8ca7f">